+++ package.json
@@ -30,9 +30,11 @@
         "@expressive-code/plugin-collapsible-sections": "^0.37.1",
         "@expressive-code/plugin-line-numbers": "^0.37.1",
         "astro": "^4.16.8",
+        "html-entities": "^2.5.2",
         "mermaid": "^11.4.0",
         "node-html-parser": "^6.1.13",
         "preact": "^10.24.3",
+        "sanitize-html": "^2.13.1",
         "tailwindcss": "^3.4.14",
         "typescript": "^5.6.3"
     },
@@ -41,6 +43,7 @@
     "license": "ISC",
     "devDependencies": {
         "@biomejs/biome": "1.9.3",
+        "@types/sanitize-html": "^2.13.0",
         "vite": "^5.4.10"
     }
 }
