+++ src/styles/over-w-props.css
@@ -1,9 +1,9 @@
 /* Dark mode colors. */
+/* https://github.com/withastro/starlight/blob/main/packages/starlight/style/props.css */
 :root {
-    --sl-color-accent-low: #131e4f;
+    /* --sl-color-accent-low: #131e4f;
     --sl-color-accent: #e7ff34;
-    /* --sl-color-accent: #182775; */
-    --sl-color-accent-high: #fffab3;
+    --sl-color-accent-high: #fffab3; */
     --sl-color-white: #ffffff;
     --sl-color-text: #d4d7dd;
     --sl-color-gray-1: #eceef2;
@@ -17,6 +17,19 @@
     /* --sl-font: "Hannari", ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol",
         "Noto Color Emoji", ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue",
         Arial, "Noto Sans", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji"; */
+    /* Shadows (dark mode) */
+    --sl-shadow-sm: 0px 1px 1px hsla(0, 0%, 0%, 0.12),
+        0px 2px 1px hsla(0, 0%, 0%, 0.24);
+    --sl-shadow-md: 0px 8px 4px hsla(0, 0%, 0%, 0.08),
+        0px 5px 2px hsla(0, 0%, 0%, 0.08), 0px 3px 2px hsla(0, 0%, 0%, 0.12),
+        0px 1px 1px hsla(0, 0%, 0%, 0.15);
+    --sl-shadow-lg: 0px 25px 7px hsla(0, 0%, 0%, 0.03),
+        0px 16px 6px hsla(0, 0%, 0%, 0.1), 0px 9px 5px hsla(223, 13%, 10%, 0.33),
+        0px 4px 4px hsla(0, 0%, 0%, 0.75), 0px 4px 2px hsla(0, 0%, 0%, 0.25);
+
+    --inset-shadow: inset 3px 3px 10px hsl(0deg 0% 0% / 74.12%),
+        inset -3px -3px 10px hsl(0deg 0% 38.43%);
+
     --sl-sidebar-width: 18rem;
     --sl-content-width: 100%;
 
@@ -60,6 +73,17 @@
     --sl-color-gray-6: #eceef2;
     --sl-color-gray-7: #f5f6f8;
     --sl-color-black: #ffffff;
+
+    /* Shadows (light mode) */
+    --sl-shadow-sm: 0px 1px 1px hsla(0, 0%, 0%, 0.06),
+        0px 2px 1px hsla(0, 0%, 0%, 0.06);
+    --sl-shadow-md: 0px 8px 4px hsla(0, 0%, 0%, 0.03),
+        0px 5px 2px hsla(0, 0%, 0%, 0.03), 0px 3px 2px hsla(0, 0%, 0%, 0.06),
+        0px 1px 1px hsla(0, 0%, 0%, 0.06);
+    --sl-shadow-lg: 0px 25px 7px rgba(0, 0, 0, 0.01),
+        0px 16px 6px hsla(0, 0%, 0%, 0.03),
+        0px 9px 5px hsla(223, 13%, 10%, 0.08), 0px 4px 4px hsla(0, 0%, 0%, 0.16),
+        0px 4px 2px hsla(0, 0%, 0%, 0.04);
 
     & main {
         background-image: linear-gradient(
